Remove unnecessary reapplication of search styles
Originally added in https://github.com/ajaxorg/ace/commit/29d8a37790d315c684735b1341bee7e04ddd55d1
as an "IE11 fix", however we don't believe this line is necessary.
Removing it and testing via BrowserStack seems to have no effect, so
can be safely removed.

This line additionally causes a problem with Overleaf's monkey-patching
of the search html & styles. It prevents the removal of the built-in Ace
styles when monkey-patching to replace with our custom styles.